refactor: reorganize component structure; update routing in App.jsx; update styles for consistency and improve layout handling in screens
