--- src/components/BottomTab.jsx
@@ -1,2 +1,0 @@
-// Compatibility export: forward to new layout component
-export { default } from './layout/BottomTab';
