+++ src/screens/ProfileScreen.jsx
@@ -130,19 +130,19 @@ const ProfileScreen = () => {
 
 const styles = {
   container: {
-    backgroundColor: theme.colors.oscarGray,
-    height: "100vh",
     display: "flex",
     flexDirection: "column",
+    height: "100vh",
+    background: theme.colors.oscarGray,
     fontFamily: theme.font.family,
+    overflow: "hidden",
   },
   content: {
     flex: 1,
-    padding: "20px",
-    maxWidth: "600px",
-    margin: "0 auto",
-    width: "100%",
-    boxSizing: "border-box",
+    overflowY: "auto",
+    padding: "16px",
+    display: "flex",
+    flexDirection: "column",
   },
 };
